+++ examples/typescript/src/client.ts
@@ -4,6 +4,9 @@ import Client, {
   SubscribeRequest,
   SubscribeRequestFilterAccountsFilter,
   SubscribeRequestFilterAccountsFilterLamports,
+  SubscribeUpdateTransactionInfo,
+  encode,
+  WasmUiTransactionEncoding,
 } from "@triton-one/yellowstone-grpc";
 
 async function main() {
@@ -83,6 +86,15 @@ async function subscribeCommand(client, args) {
 
   // Handle updates
   stream.on("data", (data) => {
+    if (data.transaction && args.transactionsParsed) {
+      const encoded = SubscribeUpdateTransactionInfo.encode(
+        data.transaction.transaction
+      ).finish();
+      const tx = encode(encoded, WasmUiTransactionEncoding.Json, 255, true);
+      console.log(`TX filters: ${data.filters}, slot#${data.transaction.slot}, tx: ${tx}`);
+      return;
+    }
+
     console.log("data", data);
   });
 
@@ -373,6 +385,11 @@ function parseCommandLineArgs() {
           default: [],
           description: "filter required account in transactions",
           type: "array",
+        },
+        "transactions-parsed": {
+          default: false,
+          describe: "parse transaction to json",
+          type: "boolean",
         },
         "transactions-status": {
           default: false,